requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
reportlab==4.3.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
